feat(graphviz): split graphviz package into core and react packages
- remove React code from @thingco/graphviz and move code from core/ directory to root src/ directory.
- create @thingco/graphviz-web and move React code to there, specify @thingco/graphviz as a peer dependency.

The reason for this is that it is difficult to build a package that exports both React Native and React components: this leads to dependency issues and problems with builds.Splitting into seperate packages removes this issue at the cost of slightly more complexity in terms of builds. The tradeoff is worth it.