feat(auth-flows): remove @xstate/inspect as direct dependency
inspect, whilst very useful

a. doesn't work on RN, so needs workarounds (checks for environment before initialising)
b. these workarounds meant that the auth system was being initialised twice

Much simpler is to just not include it: the `inspect` functionality works via
messages being passed from the machine, so much easier than packaging it into the
library is to just call it where it is being used (in a web testing context).

Remove from package, rebuild and bump version to rc.5