feat(authentication): hooks for all the login flow states
- Add one hook per [exposed] login flow state (OTP/username & password)
- Add tests for all of these

Each login flow state has a corresponding hook. In turn, each hook takes
a callback function (eg for useCheckSession, it would use [IRL] Amplify's
Auth.currentSession() function).

The hook handles firing the callback and sending the correct events to the
authentication system, exposing a function to do that + isLoading, isActive
flags + any additional functions to send additional events (for example,
requesting a new password).

The hook tests run through every one of these hooks, and check whether

1. every state that should be exposed to users as defined in the actual
   authentication system is covered with a hook, and
2. that, given a callback of the expected shape, the correct events are
   emitted by the hook.

This took quite a bit of mocking, as the hook tests should not care **at all**
about the authentication system, just that they emit expected events.

Additionally, the typings for the ohooks attempt to add root-level confidence
that everything that needs to be covered is covered, and that they are named
appropriately.