test()authentication): model testing for login flows
The model tests for authentication attempt to simulate a user interacting with the system.

The two tests runs added here (plus the modifications made to the code as the test were built, to deal with failing
cases) are for an OTP flow and a username/password flow. There are still some issues with
the correct error not showing during one specific flow, but overall there should be some
confidence that these flows are at least functional.

Both runs use an XState machine to describe the System Under Test, and are quite explicit
about what a user should be seeing -- each state describes one of these situations. The
events simulate them clicking/typing things and firing off responses -- for every one
of these events, we fire an actual event at the actual authentication system.

**All they do is check that every state has a path to it**, they do not test any other
functionality. They are designed to ensure that there are no missing routes to actual
states in the actual authentication system, that a user does not get stuck on certain screens.