fix(authentication): tighten logic for remaining OTP password tries
Move "setAttemptsMade" useState setter call in the "useAwaitingOtpInput" hook into the catch block. Currently, it is in the
finally block, where it sets the attemptsMade value to attemptsMade + 1. Which means that the attemptsMade will
keep increasing beyone the maximum attempts value. When the value hits 3, then it will bump the user back to the
awaitingOtpUsernameInput state, but otherwise it will always just say that the password was incorrect, along with 3 - attemptsMade
as the remaining tries (this gets ridiculous, as it goes negative quite quickly).

By setting the attemptsMade to 0 when the current attempts go past the limit, otherwise setting it to 3 - attemptsMade,  this fixes the issue.