fix(authentication): rename state IDs & exports
Fix a slight annoyance when building the hooks and selectors: they are
a bit verbose so have been renamed, and they also aren't amenable to search/replace. By
calitalising the state IDs (*ie* making the PascalCase), very easy
to search/replace everything in the library.

For exports, realised immediately that exporting under a namespace of
"Auth" conflicets with the Amplify's "Auth" module. So for hooks that
are used when a user is authenticated, just export them bare -- with the
renaming, this works.

Also, bump version to 0.7.0