build: remove husky, add commitizen, add .git-hooks dir + setup script, add minor change to auth
Husky (the node githooks library) seems somewhat pointless. Linting commit messages, not so much. So
remove Husky + the githooks it created, add and set up commitizen, create a .git-hooks directory and
alter the git config to point to it, add the recommended commitizen script to it to have commitizen
run on commit.

A minor change has been made to the authentication library -- the two new config values
(allowedOtpRetries and pinLength) are now available from the useAuthentication hook. This is
non-breaking and is being used as a test of commitizen.