feat(authentication): add a flag to auth system context to denote completed login
If a user has a session, but that session has expired, they will already have a PIN set.
In that case, they need to go through the authentication process but they should not need
then need to go through the PIN process. So a flag is set to false when the session is
first checked, then if the user goes through the authentication process, it is set to true.
When the system checks for an existing PIN, if the flag is true and there is an existing PIN,
skip straight to authenticated.

Originally, this commit attempted to alter a lot more of the login flow to deal with expired
refresh tokens, but this work was aborted. However, many of the ancilliary amendments have
been left in, apologies for the size.

- yarn + plugins updated.
- updates made to outdated packages
- all hook test files bar smoke tests removed: these can be tested via the model-based integration tests.
- CheckingForSession renamed to CheckingSession (BREAKING CHANGE)
- Playground app deleted
- scripts amended to remove build step for authentication & logger libs: these are now just packaged .ts files