refactor(authentication): split library into core & react parts
Shrink the scope of each part of the library by splitting it
into a core part, which has main authentication system logic,
and a React part, which has the bindings for it. So:

- @thingco/authentication-core
- @thingco/authentication-react

Remove the build step, as can just tarball the TS files as the
library; NPM seems to have no issue doing this.

**For JS/Node packages: **

Remove Jest and replace with UVU + an esbuild runner, which
speeds up the tests by a significant amount (approximately
an order of magnitude reduction in time).

This creates an issue w/r/t mocks/spies for the API calls,

So for React based stuff that needs mocking of deep dependencies,
Jest is used despite being extremely slow.

Fix some of the tests in core: moving them turned up a few
test issues.