fix(authentication-react): add ref to each hook to track mounted state
In the hooks, what can happen is that the async callback fires, the callback receives
a response, which then triggers a message to the state machine, which triggers a state
change. The loading flag changes to true when the request begins, then switches to false
after it ends. However, at that point in time, state transition may already have occurred.
This will cause a memory leak. So to fix this, all hooks that make async requests & require
an isLoading flag have had a ref called isMounted added to them. They also all have a useEffect,
which, in its returned callback function, sets isMounted to false. Then, in the async function,
the setIsLoading function will only fire if isMounted's current value is true.

Version bump made (patch).