remove the average speed formatter
The "average speed" formatter works slightly differently to the rest of the formatters work. Instead of just accepting a single value & returning a formatted value, it takes two values, runs a calculation then formats the resulting value as speed.

So created an "averageSpeedInKmph" function in caclulators.ts; that can be used to produce a speed value, and delete the average speed formatter ("speed" should be used instead).