chore: round of renaming
This commit is the start of getting this repo into a sensible shape, building upon lessons learnt
from building out the authentication and the newer shell of the data-transformers library.

NOTE that this doesn't change how anything works. The renamed folders still have the same
actual names in their package.json files, so publishing should be completely unaffected. And
the existing versions are still available from ThingCo's GH packages, so everything will
continue to work exactly as it does atm.

- @thingco/auth-flows needs to be killed, the `@thingco/authentication-{core|react}` libs
  obselete it.
- `@thingco/data-transformers-core` should contain all base logic from `@thingco/unit-formatter` and
  `@thingco/graphviz`. The data transformers need to part of the functional core of the application;
  (note that currently they are stateful, this probably needs to be changed).
- `@thingco/shared-types` shouldn't need to exist. It currently does, but slapping a deprecation notice
  on the folder should make it obvious it's going to be killed.
- `@thingco/user-preferences` shouldn't be standalone; `@thingco/user-preferences` & the supporting
  persistence layer are the next to go, so that can be killed as well.

Removed all scripts from root; no longer really needed. The library are published as Typescript
and do not require a build step.  Mostly of the remaining scripts are just aliases, and the aliases go
out of date, so nix them. Because no longer any global scripts, the `./scripts` folder + the
dependency on `zx` can both be excised

The component libraries need serious work, so freeze dev on the existing ones (mark as deprecated as
well) bar hotfixes as get things into place.

Also, create an `apps` directory at root. This is a place to build out and dogfood the other libs:
going forward move (for example) the theme builder app into here.